+++ workspaces/azure-devops/plugins/search-backend-module-azure-devops/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-search-backend-module-azure-devops
 
+## 0.9.0
+
+### Minor Changes
+
+- 6c8035e: Backstage version bump to v1.54.5
+
 ## 0.8.0
 
 ### Minor Changes